+++ apps/studio/components/interfaces/Database/Replication/BatchRestartDialog.tsx
@@ -15,9 +15,9 @@ import {
 import { PipelineStatusName } from './Replication.constants'
 import { RestartCostEstimate } from './RestartCostEstimate'
 import { getTableCopyTargets } from './TableSyncCopy.utils'
-import type { TableSyncCopyConfig } from '@/data/replication/create-destination-pipeline-mutation'
 import { ReplicationPipelineTableStatus } from '@/data/replication/pipeline-replication-status-query'
 import { useRollbackTablesMutation } from '@/data/replication/rollback-tables-mutation'
+import type { TableSyncCopyConfig } from '@/data/replication/types'
 
 interface BatchRestartDialogProps {
   open: boolean
