+++ apps/studio/data/replication/types.ts
@@ -1,3 +1,33 @@
+import { components } from 'api-types'
+
+type CreateDestinationPipelineBody =
+  components['schemas']['CreateReplicationDestinationPipelineBody']
+export type CreatePipelineApiConfig = CreateDestinationPipelineBody['pipeline_config']
+
+export type BatchConfig = {
+  maxFillMs?: number
+  maxBytes?: number
+  memoryBudgetRatio?: number
+}
+
+export type TableSyncCopyConfig = NonNullable<CreatePipelineApiConfig['table_sync_copy']>
+
+export type PipelineConfig = {
+  publicationName: string
+  batch?: BatchConfig
+  maxTableSyncWorkers?: number
+  maxCopyConnectionsPerTable?: number
+  invalidatedSlotBehavior?: 'error' | 'recreate'
+  tableSyncCopy: TableSyncCopyConfig
+}
+
+export type DestinationConfig =
+  | { bigQuery: BigQueryDestinationConfig }
+  | { iceberg: IcebergDestinationConfig }
+  | { ducklake: DucklakeDestinationConfig }
+  | { snowflake: SnowflakeDestinationConfig }
+  | { clickHouse: ClickHouseDestinationConfig }
+
 // "Custom parameters" DuckLake: caller provides the PostgreSQL catalog URL and the
 // S3-compatible storage credentials directly.
 export type DucklakeManualDestinationConfig = {
@@ -27,3 +57,39 @@ export type DucklakeSupabaseDestinationConfig = {
 export type DucklakeDestinationConfig =
   | DucklakeManualDestinationConfig
   | DucklakeSupabaseDestinationConfig
+
+export type BigQueryDestinationConfig = {
+  projectId: string
+  datasetId: string
+  serviceAccountKey: string
+  connectionPoolSize?: number
+  maxStalenessMins?: number
+}
+
+export type IcebergDestinationConfig = {
+  projectRef: string
+  warehouseName: string
+  namespace?: string
+  catalogToken: string
+  s3AccessKeyId: string
+  s3SecretAccessKey: string
+  s3Region: string
+}
+
+export type SnowflakeDestinationConfig = {
+  accountId: string
+  user: string
+  privateKey: string
+  privateKeyPassphrase?: string
+  database: string
+  schema: string
+  role?: string
+}
+
+export type ClickHouseDestinationConfig = {
+  url: string
+  user: string
+  password?: string
+  database: string
+  engine?: 'merge_tree' | 'replacing_merge_tree'
+}
