+++ apps/studio/data/replication/utils.ts
@@ -1,4 +1,9 @@
-import type { DucklakeDestinationConfig, DucklakeSupabaseDestinationConfig } from './types'
+import type {
+  CreatePipelineApiConfig,
+  DucklakeDestinationConfig,
+  DucklakeSupabaseDestinationConfig,
+  PipelineConfig,
+} from './types'
 import { MAX_RETRY_FAILURE_COUNT } from '@/data/query-client'
 import { ResponseError } from '@/types'
 
@@ -43,3 +48,25 @@ export function isDucklakeSupabaseConfig(
 ): config is DucklakeSupabaseDestinationConfig {
   return 'catalogProjectRef' in config
 }
+
+export const buildPipelineApiConfig = ({
+  publicationName,
+  batch,
+  maxTableSyncWorkers,
+  maxCopyConnectionsPerTable,
+  invalidatedSlotBehavior,
+  tableSyncCopy,
+}: PipelineConfig): CreatePipelineApiConfig => ({
+  publication_name: publicationName,
+  max_table_sync_workers: maxTableSyncWorkers,
+  max_copy_connections_per_table: maxCopyConnectionsPerTable,
+  invalidated_slot_behavior: invalidatedSlotBehavior,
+  table_sync_copy: tableSyncCopy,
+  batch: batch
+    ? {
+        max_fill_ms: batch.maxFillMs,
+        max_bytes: batch.maxBytes,
+        memory_budget_ratio: batch.memoryBudgetRatio,
+      }
+    : undefined,
+})
